Ran 2to3 on the whole project which made a few small modifications. Fixed a couple of things that 2to3 broke related to urllib. Changed gulp macros to invoke python3 instead of python. Changed a couple of shebangs to point to python3.

+++ main/config.py
@@ -19,7 +19,7 @@ else:
 
   CURRENT_VERSION_ID = os.environ.get('CURRENT_VERSION_ID')
   CURRENT_VERSION_NAME = CURRENT_VERSION_ID.split('.')[0]
-  CURRENT_VERSION_TIMESTAMP = long(CURRENT_VERSION_ID.split('.')[1]) >> 28
+  CURRENT_VERSION_TIMESTAMP = int(CURRENT_VERSION_ID.split('.')[1]) >> 28
   if DEVELOPMENT:
     import calendar
